fix: properly set post conditions from payload
